Rename "lithOps" to "lithium_helpers".
